requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.1.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
distlib==0.3.1
filelock==3.0.12
google-api-core==1.22.1
google-api-python-client==1.10.1
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
multidict==4.7.6
oauthlib==3.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.31
yarl==1.5.1
